change to float32

+++ garment_adapter/garment_diffusion.py
@@ -73,7 +73,7 @@ class ClothAdapter:
 
         cloth = prepare_image(cloth_image, height, width)
         cloth_mask = prepare_mask(cloth_mask_image, height, width)
-        cloth = (cloth * cloth_mask).to(self.device, dtype=torch.float16)
+        cloth = (cloth * cloth_mask).to(self.device, dtype=torch.float32)
 
         if prompt is None:
             prompt = "a photography of a model"
@@ -140,7 +140,7 @@ class ClothAdapter:
 
         cloth = prepare_image(cloth_image, height, width)
         cloth_mask = prepare_mask(cloth_mask_image, height, width)
-        cloth = (cloth * cloth_mask).to(self.device, dtype=torch.float16)
+        cloth = (cloth * cloth_mask).to(self.device, dtype=torch.float32)
 
         with torch.inference_mode():
             prompt_embeds_null = self.pipe.encode_prompt([""], device=self.device, num_images_per_prompt=num_images_per_prompt, do_classifier_free_guidance=False)[0]
@@ -242,7 +242,7 @@ class ClothAdapter_AnimateDiff:
 
         cloth = prepare_image(cloth_image, height, width)
         cloth_mask = prepare_mask(cloth_mask_image, height, width)
-        cloth = (cloth * cloth_mask).to(self.device, dtype=torch.float16)
+        cloth = (cloth * cloth_mask).to(self.device, dtype=torch.float32)
 
         if prompt is None:
             prompt = "a asian girl with big smile"
